+++ frontend/src/components/versions/index.ts
@@ -1,4 +1,5 @@
 export { EtcdSystemdModal } from './EtcdSystemdModal';
+export { KubeadmCertsModal } from './KubeadmCertsModal';
 export { KernelParamsCollectModal } from './KernelParamsCollectModal';
 export { NodeNicsCollectModal } from './NodeNicsCollectModal';
 export { KubeletConfigCollectModal } from './KubeletConfigCollectModal';